refactor: move Injury back to model person package
